+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Use official markdownlint github action
 - Use variable debug format in debug logs
 - Update command include and exclude lists
+- Print included contexts in red
 
 ## v2.3.0